implement check status

+++ src/main/java/com/hackathon/enrollmentSystem/enrollmentSystem/entity/Application.java
@@ -35,7 +35,7 @@ public class Application extends IdBasedEntity {
     @OneToOne
     private Course course;
 
-    private Long Tracking;
+    private Integer Tracking;
 
 
 }
